backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests

aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
autocommand==2.2.2
Babel==2.14.0
bcc==0.30.0
bcrypt==4.2.0
blinker==1.8.2
Brlapi==0.8.5
certifi==2024.6.2
chardet==5.2.0
click==8.1.7
cloud-init==24.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==42.0.5
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.9
duplicity==2.1.4
fasteners==0.18
frozenlist==1.4.1
gyp-next==0.16.2
httplib2==0.22.0
idna==3.6
importlib_metadata==8.2.0
inflect==7.3.1
jaraco.classes==3.4.0
jaraco.context==6.0.0
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==3.1.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
keyring==25.3.0
language-selector==0.1
launchpadlib==2.0.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.30.0
Mako==1.3.5.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
more-itertools==10.3.0
multidict==6.0.5
netaddr==0.10.1
oauthlib==3.2.2
olefile==0.46
packaging==24.1
paramiko==3.4.0
pexpect==4.9.0
pillow==10.4.0
proton-core==0.3.3
proton-keyring-linux==0.1.0
proton-vpn-api-core==0.35.5
proton-vpn-gtk-app==4.6.0
proton-vpn-network-manager==0.9.1
ptyprocess==0.7.0
pycairo==1.26.1
pycares==4.4.0
pycups==2.0.4
Pygments==2.18.0
PyGObject==3.48.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyserial==3.5
python-apt==2.9.0+ubuntu1
python-dateutil==2.9.0
python-debian==0.1.49+ubuntu3
python-gnupg==0.5.2
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.3
sentry-sdk==1.40.4
six==1.16.0
systemd-python==235
typeguard==4.3.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.16
wadllib==1.3.6
wheel==0.44.0
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==3.20.0
